requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0

psycopg2-binary==2.9.9  
SQLAlchemy==2.0.25

bcrypt==4.0.1  
cryptography==41.0.7  
python-dotenv==1.0.0 

requests==2.31.0  
python-dateutil==2.8.2  
PyJWT==2.8.0  